test(release): isolate provenance verification coverage
